task_app/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
rest-framework-simplejwt==0.0.2
setuptools==72.1.0
sqlparse==0.5.1
wheel==0.43.0
